requirements.txt
aiogram==2.25.1
pytz==2023.3
sqlalchemy==2.0.12
alembic==1.10.4
python-dotenv==1.0.0
APScheduler==3.10.1 